refactor: participantCount jpa 메서드로 처리

+++ src/main/kotlin/com/yapp/lettie/api/timecapsule/service/dto/TimeCapsuleDetailDto.kt
@@ -8,7 +8,7 @@ data class TimeCapsuleDetailDto(
     val title: String,
     val subtitle: String?,
     val openAt: LocalDateTime,
-    val participantCount: Int,
+    val participantCount: Long,
     val likeCount: Long,
     val isLiked: Boolean? = null,
     val status: TimeCapsuleStatus,
